+++ nym-vpn-app/src/screens/node/Node.tsx
@@ -1,72 +1,45 @@
-import { useEffect, useState } from 'react';
 import { useNavigate } from 'react-router';
 import { Trans, useTranslation } from 'react-i18next';
 import { motion } from 'motion/react';
 import {
-  UiCountry,
+  SelectedUiNode,
   UiGateway,
-  UiGatewaysByCountry,
   useDialog,
   useMainDispatch,
   useMainState,
   useNodeList,
   useNodeListState,
 } from '../../contexts';
-import { NodeHop, StateDispatch, isGateway } from '../../types';
+import { NodeHop, StateDispatch } from '../../types';
 import { Link, PageAnim, TextInput } from '../../ui';
 import { kvSet } from '../../kvStore';
-import { uiNodeToRaw } from '../../contexts/node-list/util';
+import { uiNodeToSelectedNode } from '../../contexts/node-list/util';
 import { useI18nError } from '../../hooks';
 import { routes } from '../../router';
 import LocationDetailsDialog from './LocationDetailsDialog';
-import { NodeList } from './list';
+import { NodeList, useFilterList } from './list';
 
 function Node({ node }: { node: NodeHop }) {
   const { backendFlags, vpnMode, quic } = useMainState();
   const dispatch = useMainDispatch() as StateDispatch;
 
   const { isOpen, close } = useDialog();
-  const { nodes, loading, gateways, error } = useNodeList();
+  const { loading, error } = useNodeList();
   const { setFocused, reset: resetSaved, addToExpanded } = useNodeListState();
   const { tE } = useI18nError();
 
-  const [uiNodes, setUiNodes] = useState<UiGatewaysByCountry[]>(nodes);
-  const [uiGateways, setUiGateways] = useState<UiGateway[]>(gateways);
-  const [search, setSearch] = useState('');
   const quicFilter =
     vpnMode === 'wg' && node === 'entry' && backendFlags.quic && quic;
 
   const navigate = useNavigate();
   const { t } = useTranslation('nodeLocation');
 
-  // refresh the UI list whenever the backend gateway data changes
-  useEffect(() => {
-    setUiNodes(nodes);
-    setUiGateways([]);
-    setSearch('');
-  }, [nodes, gateways]);
+  const { filter, nodes, gateways } = useFilterList();
 
-  const filter = (value: string) => {
-    if (value.length > 0) {
-      const filteredNodes = nodes.filter((node) => {
-        // toLowerCase() is used to make it case-insensitive
-        return node.i18n.toLowerCase().includes(value.toLowerCase());
-      });
-      const filteredGw = gateways.filter((gw) => {
-        return gw.name.toLowerCase().includes(value.toLowerCase());
-      });
-      setUiNodes(filteredNodes);
-      setUiGateways(filteredGw);
-    } else {
-      setUiNodes(nodes);
-      setUiGateways([]);
-    }
-    setSearch(value);
-  };
-
-  const handleSelect = async (selected: UiCountry | UiGateway) => {
+  const handleSelect = async (selected: SelectedUiNode) => {
+    const selectedNode = uiNodeToSelectedNode(selected);
     if (
-      isGateway(selected) &&
+      selectedNode.type === 'gateway' &&
       (selected.isSelected === 'exit' || selected.isSelected === 'entry')
     ) {
       return;
@@ -74,11 +47,11 @@ function Node({ node }: { node: NodeHop }) {
 
     await kvSet(
       node === 'entry' ? 'entry-node' : 'exit-node',
-      uiNodeToRaw(selected),
+      uiNodeToSelectedNode(selected),
     );
     dispatch({
       type: 'set-node',
-      payload: { hop: node, node: selected },
+      payload: { hop: node, node: selectedNode },
     });
     navigate(routes.root);
     resetSaved(node);
@@ -154,7 +127,7 @@ function Node({ node }: { node: NodeHop }) {
             </p>
           )}
           <TextInput
-            value={search}
+            defaultValue=""
             onChange={filter}
             placeholder={t('search-country')}
             leftIcon="search"
@@ -175,11 +148,11 @@ function Node({ node }: { node: NodeHop }) {
         )}
         {!loading && (
           <NodeList
-            nodes={uiNodes}
-            gateways={uiGateways}
+            nodes={nodes}
+            gateways={gateways}
             onSelect={handleSelect}
             onNodeDetails={handleNodeDetails}
-            node={node}
+            hop={node}
             vpnMode={vpnMode}
           />
         )}