+++ nym-vpn-app/src/screens/node/list/index.ts
@@ -1,1 +1,2 @@
 export { default as NodeList } from './NodeList';
+export { useFilterList } from './useFilterList';