+++ nym-vpn-core/crates/nym-vpn-api-client/src/request.rs
@@ -4,7 +4,15 @@
 use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Serialize, Deserialize)]
-pub struct CreateAccountRequestBody {
+pub struct CreateAndroidAccountRequestBody {
+    pub account_addr: String,
+    pub pub_key: String,
+    pub signature_base64: String,
+    pub purchase_token: String,
+}
+
+#[derive(Debug, Serialize, Deserialize)]
+pub struct CreateAppleAccountRequestBody {
     pub account_addr: String,
     pub pub_key: String,
     pub signature_base64: String,