+++ nym-vpn-core/crates/nym-vpn-api-client/src/types/platform.rs
@@ -1,14 +1,8 @@
 // Copyright 2025 - Nym Technologies SA <contact@nymtech.net>
 // SPDX-License-Identifier: GPL-3.0-only
 
-#[derive(Clone, Copy, Debug)]
+#[derive(Clone, Debug)]
 pub enum Platform {
     Apple,
-}
-impl Platform {
-    pub fn api_path_component(&self) -> &'static str {
-        match self {
-            Platform::Apple => crate::routes::APPLE,
-        }
-    }
+    Android { purchase_token: String },
 }